+++ promptlauncher/__init__.py
@@ -1,0 +1,13 @@
+from .gui import PromptWindow
+from .tray import create_tray
+from .hotkey import get_custom_hotkey
+from .ssh_backup import SshBackupManager
+from .main import main
+
+__all__ = [
+    "PromptWindow",
+    "create_tray",
+    "get_custom_hotkey",
+    "SshBackupManager",
+    "main",
+]
